requirements.txt

python==3.9.13
streamlit==1.30.0
requests==2.31.0
python-docx==1.1.0
Pillow==10.2.0
PyGithub==2.1.1
pandas==2.1.4
